No longer use Geocode for UPS Timestamp
UPS has a request flag that allows the returning of GMT time information
which can be used directly without the need for geocode tools to help
with the timezone conversion logic.

This commit only commented out the existing code to use the GMT time
instead of geocode logic. It is feasible to thing that some mode may
also be obsolete as a result and more cleanups can take place.