Apply a "just in case" default timezone
Since timezone should be UTC for the GMTTime approach on UPS and the
non-GMTTime should be ET, I updated the default timezone for when an
address does not have a timezone correction to more appropriately match
the multi-format logic I put in place.

No noticable differences happened in any examples I had but in theory
this would be a bug and so I applied the theoretical fix for it.