Allow an option to include the status information for UPS
As a user, I'd like to process the status information found
on the activity of the UPS information, as this is more reliable
than relying on the human readable description; which may change
or be altered.  The status codes are stable and good indicators
of the true state of the package.

 On branch ups-status
 Changes to be committed:
        modified:   carriers/ups.js
        modified:   index.js